requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
pydantic>=2.4.2
python-pptx>=0.6.21
python-docx>=0.8.11
httpx>=0.25.0 